requirements/base.txt
gpiozero
PyYAML
